+++ js/jquery.feedback_me.js
@@ -4,7 +4,7 @@
 * jQuery Feedback Plugin
 * 
 * File:        jquery.feedback_me.js
-* Version:     0.2.5
+* Version:     0.2.7
 * Author:      Daniel Reznick
 * Info:        https://github.com/vedmack/feedback_me
 * Contact:     vedmack@gmail.com	
@@ -26,13 +26,13 @@
 				Required:			true
 				Type:				String
 				Description:		URL of your servlet/php etc ('name', 'message' and 'email' parameters will be send to your servlet/php etc...)
-				
+
 * jQueryUI
 				Required:			false
 				Type:				boolean
 				Default value:		false
 				Description:		Tell the plugin to use jQuery UI theme
-					
+
 * bootstrap
 				Required:			false
 				Type:				boolean
@@ -44,7 +44,7 @@
 				Type:				boolean
 				Default value:		false
 				Description:		Tell the plugin to display email input field
-										
+
 * name_label
 				Required:			false
 				Type:				String
@@ -56,13 +56,13 @@
 				Type:				String
 				Default value:		"Email"
 				Description:		Label for email input
-		
+
 * message_label
 				Required:			false
 				Type:				String
 				Default value:		"Message"
 				Description:		Label for message input
-									
+
 * submit_label
 				Required:			false
 				Type:				String
@@ -74,13 +74,13 @@
 				Type:				String
 				Default value:		"Feedback"
 				Description:		Label for title text
-		
+
 * trigger_label
 				Required:			false
 				Type:				String
 				Default value:		"Feedback"
 				Description:		Label for open/close (trigger) button
-		
+
 * name_placeholder
 				Required:			false
 				Type:				String
@@ -92,12 +92,36 @@
 				Type:				String
 				Default value:		""
 				Description:		Watermark for email input
-		
+
 * message_placeholder
 				Required:			false
 				Type:				String
 				Default value:		""
 				Description:		Watermark for message input
+
+* name_required
+				Required:			false
+				Type:				boolean
+				Default value:		false
+				Description:		Makes input required
+
+* email_required
+				Required:			false
+				Type:				boolean
+				Default value:		false
+				Description:		Makes input required
+
+* message_required
+				Required:			false
+				Type:				boolean
+				Default value:		false
+				Description:		Makes input required
+
+* show_asterisk_for_required
+				Required:			false
+				Type:				boolean
+				Default value:		false
+				Description:		Add an asterisk to the label of the required inputs
 
 * close_on_click_outisde				
 				Required:			false
@@ -127,6 +151,7 @@ var fm = (function () {
 		}
 
 		if ($("#feedback_trigger").hasClass("feedback_trigger_closed")) {
+
 			$("#feedback_trigger , #feedback_content").animate(
 				animation_show,
 				500,
@@ -137,16 +162,15 @@ var fm = (function () {
 				}
 			);
 		} else {
+			//first add the closed class so double (which will trigger closeFeedback function) click wont try to hide the form twice
+			$("#feedback_trigger").addClass("feedback_trigger_closed");
+			$("#feedback_content").addClass("feedback_content_closed");
+
 			$("#feedback_trigger , #feedback_content").animate(
 				animation_hide,
-				500,
-				function () {
-					$("#feedback_trigger").addClass("feedback_trigger_closed");
-					$("#feedback_content").addClass("feedback_content_closed");
-				}
+				500
 			);
 		}
-
 	}
 
 	function closeFeedback(event) {
@@ -170,6 +194,36 @@ var fm = (function () {
 		);
 	}
 
+	function validateFeedbackForm() {
+		if ((fm_options.name_required === true && $("#feedback_name").val() === "") ||
+				(fm_options.email_required === true && $("#feedback_email").val() === "") ||
+				(fm_options.message_required === true && $("#feedback_message").val() === "")) {
+			return false;
+		}
+		return true;
+
+	}
+
+	function checkRequiredFieldsOk() {
+		var $reqFields = $("[required]"),
+			$fm_form,
+			form_valid = true;
+
+		if ($reqFields.length > 0) {
+			//prevent form submit (needed for validation)
+			$('#feedback_me_form').submit(function (event) {
+				event.preventDefault();
+			});
+
+			$fm_form = $('#feedback_me_form');
+			form_valid = validateFeedbackForm();
+			if (form_valid === false) {
+				$fm_form.submit();
+			}
+		}
+		return form_valid;
+	}
+
 	function applyCloseOnClickOutside() {
 		if (parseFloat($().jquery) >= 1.7) {
 			$(document).on("click", document, function (event) {
@@ -182,7 +236,7 @@ var fm = (function () {
 		}
 	}
 
-	function appendFeedbackToBody(fm_options) {
+	function appendFeedbackToBody() {
 		var jQueryUIClasses1 = "",
 			jQueryUIClasses2 = "",
 			jQueryUIClasses3 = "",
@@ -193,7 +247,15 @@ var fm = (function () {
 			jquery_class = "",
 			bootstrap_class = "",
 			bootstrap_btn = "",
-			bootstrap_hero_unit = "";
+			bootstrap_hero_unit = "",
+
+			name_required = fm_options.name_required ? "required" : "",
+			email_required = fm_options.email_required ? "required" : "",
+			message_required = fm_options.message_required ? "required" : "",
+
+			name_asterisk  = fm_options.name_required && fm_options.show_asterisk_for_required ? "<span class=\"required_asterisk\">*</span>" : "",
+			email_asterisk  = fm_options.email_required && fm_options.show_asterisk_for_required ? "<span class=\"required_asterisk\">*</span>" : "",
+			message_asterisk  = fm_options.message_required && fm_options.show_asterisk_for_required ? "<span class=\"required_asterisk\">*</span>" : "";
 
 		if (fm_options.bootstrap === true) {
 			bootstrap_class = " fm_bootstrap ";
@@ -219,7 +281,7 @@ var fm = (function () {
 		}
 
 		if (fm_options.show_email === true) {
-			email_html = '<li>	<label for="feedback_email">' + fm_options.email_label + '</label> <input type="text" id="feedback_email" placeholder="' + fm_options.email_placeholder + '"></input> </li>';
+			email_html = '<li>	<label for="feedback_email">' + fm_options.email_label + '</label> ' + email_asterisk + ' <input type="text" id="feedback_email" ' + email_required + ' placeholder="' + fm_options.email_placeholder + '"></input> </li>';
 			email_feedback_content_class = "email_present";
 		}
 
@@ -231,14 +293,16 @@ var fm = (function () {
 							+ '<div class="feedback_title ' + jQueryUIClasses1 + jQueryUIClasses3 + '">'
 							+	'<span class="' + jQueryUIClasses4 + '">' + fm_options.title_label + '</span>'
 							+ '</div>'
+							+	'<form id="feedback_me_form">'
 							+	'<ul>'
-							+		'<li>	<label for="feedback_name">' + fm_options.name_label + '</label> <input type="text" id="feedback_name" placeholder="' + fm_options.name_placeholder + '"></input> </li>'
+							+		'<li>	<label for="feedback_name">' + fm_options.name_label + '</label> ' + name_asterisk + ' <input type="text" id="feedback_name" ' + name_required + ' placeholder="' + fm_options.name_placeholder + '"></input> </li>'
 
 							+		 email_html
 
-							+		'<li>	<label for="feedback_message">' + fm_options.message_label + '</label> <textarea rows="5" id="feedback_message" placeholder="' + fm_options.message_placeholder + '"></textarea> </li>'
-							+		'<li>	<button id="feedback_submit" onclick="fm.sendFeedback();" class="' + bootstrap_btn + '">' + fm_options.submit_label + '</button> </li>'
+							+		'<li>	<label for="feedback_message">' + fm_options.message_label + '</label> ' + message_asterisk + ' <textarea rows="5" id="feedback_message" ' + message_required + ' placeholder="' + fm_options.message_placeholder + '"></textarea> </li>'
+							+		'<li>	<button id="feedback_submit" onclick="fm.sendFeedback(event);" class="' + bootstrap_btn + '">' + fm_options.submit_label + '</button> </li>'
 							+	'</ul>'
+							+	'</form>'
 						+ '</div>');
 
 		if (fm_options.jQueryUI === true) {
@@ -255,7 +319,13 @@ var fm = (function () {
 
 	}
 
-	function sendFeedback() {
+	function sendFeedback(event) {
+		var checkValid = checkRequiredFieldsOk();
+		if (checkValid === false) {
+			event.cancelBubble = true;
+			event.stopPropagation();
+			return;
+		}
 		$.ajax({
 			type: 'POST',
 			url: fm.getFmOptions().feedback_url,
@@ -302,6 +372,10 @@ var fm = (function () {
 			name_placeholder : "",
 			email_placeholder : "",
 			message_placeholder : "",
+			name_required : false,
+			email_required : false,
+			message_required : false,
+			show_asterisk_for_required : false,
 			submit_label : "Send",
 			title_label : "Feedback",
 			trigger_label : "Feedback"
@@ -309,7 +383,7 @@ var fm = (function () {
 
 		fm_options = $.extend(default_options, options);
 
-		appendFeedbackToBody(fm_options);
+		appendFeedbackToBody();
 	}
 
     return {